Support overriding the text shown when option is selected
Normally, when an option is selected, the option's text is displayed in
the main control.  This commit adds an option 'generate_selected_value'
whose value should be a function that receives the selected option as an
argument and should return either { text: <string> } or { html: <string> }
as the value to be displayed.