Support overriding an option's text / HTML value
If present, use the 'data-text'/'data-html' attribute values instead of
the option's actual text/innerHTML content.  This allows an option's
displayed value to be styled arbitrarily when the select is converted to
a Chosen widget, while leaving the option's text as something that makes
sense in a plain select widget.