Increase space around the "single deselect" control
If the width of the main control is set to "auto" and deselection is
enabled, the deselect control (an "X") ends up right next to the selected
option's text in the main control.  With this adjustment, there's roughly
the same amount of space between the "X" and the text as there is between
the "X" and the downwards arrow to its right.